# Improve GetConnection to avoid allocating a temporary slice