`GetConnection` was creating a temporary slice with all active connections, then picking one of these randomly. Instead, we can iterate through the connections starting at a random index and pick the first active connection we find.