@prashantv 2016-03-05
Updated, the benchcmp output,
```
> benchcmp bench1.txt bench2.txt
benchmark                           old ns/op     new ns/op     delta
BenchmarkGetConnection0In1Out-8     131           86.2          -34.20%
BenchmarkGetConnection1In0Out-8     122           84.3          -30.90%
BenchmarkGetConnection5In5Out-8     485           93.8          -80.66%

benchmark                           old allocs     new allocs     delta
BenchmarkGetConnection0In1Out-8     1              0              -100.00%
BenchmarkGetConnection1In0Out-8     1              0              -100.00%
BenchmarkGetConnection5In5Out-8     4              0              -100.00%

benchmark                           old bytes     new bytes     delta
BenchmarkGetConnection0In1Out-8     8             0             -100.00%
BenchmarkGetConnection1In0Out-8     8             0             -100.00%
BenchmarkGetConnection5In5Out-8     120           0             -100.00%```